Updated CocoaPods instructions for latest CocoaPods
In CocoaPods 1.0.0.beta.3, the Podfile now needs to specify the target name, otherwise `pod install` fails with the following:

```[!] The dependency `RxSwift (~> 2.0)` is not used in any concrete target.
The dependency `RxCocoa (~> 2.0)` is not used in any concrete target.```

I added an extra step to the instructions to make it work again. The above format of course still works on the pre-1.0.0 cocoapods as well, so I believe it's safe to update the instructions already.